+++ lib/ruby-progressbar/base.rb
@@ -7,6 +7,7 @@ class ProgressBar
 
     def initialize(options = {})
       self.output       = options[:output] || DEFAULT_OUTPUT_STREAM
+      autostart         = options.fetch(:autostart, true)
 
       super(options)
 
@@ -15,7 +16,7 @@ class ProgressBar
       @elapsed_time     = Components::ElapsedTimer.new
       @throttle         = Components::Throttle.new(options)
 
-      start :at => options[:starting_at]
+      start :at => options[:starting_at] if autostart
     end
 
     ###
@@ -53,19 +54,19 @@ class ProgressBar
     # Stopping The Bar
     #
     def finish
-      with_update { with_progressables(:finish) }
+      with_update { with_progressables(:finish) } unless finished?
     end
 
     def pause
-      with_update { with_timers(:pause) }
+      with_update { with_timers(:pause) } unless paused?
     end
 
     def stop
-      with_update { with_timers(:stop) }
+      with_update { with_timers(:stop) } unless stopped?
     end
 
     def resume
-      with_update { with_timers(:resume) }
+      with_update { with_timers(:resume) } if stopped?
     end
 
     def reset
@@ -92,6 +93,10 @@ class ProgressBar
       with_update { @bar.progress_mark = mark }
     end
 
+    def remainder_mark=(mark)
+      with_update { @bar.remainder_mark = mark }
+    end
+
     def title=(title)
       with_update { super }
     end
@@ -100,9 +105,13 @@ class ProgressBar
     # Output
     #
     def clear
+      self.last_update_length = 0
+
       if output.tty?
         output.print clear_string
         output.print "\r"
+      else
+        output.print "\n"
       end
     end
 
@@ -128,10 +137,15 @@ class ProgressBar
     end
 
   private
-    attr_accessor   :output
+    attr_accessor   :output,
+                    :last_update_length
 
     def clear_string
       "#{" " * length}"
+    end
+
+    def last_update_length
+      @last_update_length ||= 0
     end
 
     def with_progressables(*args)
@@ -150,23 +164,37 @@ class ProgressBar
     end
 
     def update(options = {})
-      if output.tty? || stopped?
-        with_timers(:stop) if finished?
+      with_timers(:stop) if finished?
 
-        @throttle.choke( stopped? || options[:force] ) do
-          if length_changed?
-            clear
-            reset_length
-          end
+      if length_changed?
+        clear
+        reset_length
+      end
 
-          output.print self.to_s + eol
-          output.flush
+      @throttle.choke( stopped? || options[:force] ) do
+        if output.tty?
+          formatted_string = self.to_s
+          output_string    = formatted_string
+        else
+          formatted_string = self.to_s(DEFAULT_NON_TTY_FORMAT_STRING)
+          formatted_string = formatted_string[0...-1] unless finished?
+
+          output_string    = formatted_string[last_update_length..-1]
         end
+
+        self.last_update_length = formatted_string.length
+
+        output.print output_string + eol
+        output.flush
       end
     end
 
     def eol
-      stopped? ? "\n" : "\r"
+      if output.tty?
+        stopped? ? "\n" : "\r"
+      else
+        stopped? ? "\n" : ""
+      end
     end
   end
 end